fix: display research draft updates for author
