+++ src/pages/Research/Content/ResearchDescription.tsx
@@ -16,11 +16,12 @@ import { IModerationStatus } from 'oa-shared'
 import { trackEvent } from 'src/common/Analytics'
 import { logger } from 'src/logger'
 import { useResearchStore } from 'src/stores/Research/research.store'
-import { buildStatisticsLabel, researchStatusColour } from 'src/utils/helpers'
+import { buildStatisticsLabel } from 'src/utils/helpers'
 import { incrementViewCount } from 'src/utils/incrementViewCount'
 import { Box, Card, Divider, Flex, Heading, Text } from 'theme-ui'
 
 import { ContentAuthorTimestamp } from '../../common/ContentAuthorTimestamp/ContentAuthorTimestamp'
+import { researchStatusColour } from '../researchHelpers'
 
 import type { ITag } from 'src/models'
 import type { IResearch } from 'src/models/research.models'
